+++ packages/toolbars-new/web/src/icons/TextColorIcon.tsx
@@ -12,7 +12,11 @@ const TextColorIcon = props => (
     {...props}
   >
     <g id="Specs-for-devs" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-      <g id="Icons" transform="translate(-178.000000, -127.000000)" fill="currentColor">
+      <g
+        id="Icons"
+        transform="translate(-178.000000, -127.000000)"
+        fill="var(--ricos-settings-icons-color, #000000)"
+      >
         <g
           id="Icons-/-Formatting-/-Text-color-/-All-Copy-11"
           transform="translate(178.000000, 127.000000)"